Fixed random pattern

+++ benchmarks/latency/settings.h
@@ -18,7 +18,7 @@ struct Settings
     bool            latency;
     int             cudaDevice;
     const char*     controllerPath;
-    uint32_t        controllerId;
+    uint64_t        controllerId;
     uint32_t        adapter;
     uint32_t        segmentId;
     uint32_t        nvmNamespace;
